Create biomass at age table for MACE summer GOA cruise reports

Build the pollock biomass-at-age table used in summer GOA reports. This will report biomass at age for ages from 1- maximum reported age in the survey, and will do so for each report number (i.e. survey region) that was defined for the analysis.

Usage

build_biomass_at_age_table_summer_goa(biomass_nums_age_data)

Arguments

biomass_nums_age_data

The name of the dataframe that contains the pollock biomass-at-age data; this dataframe is created in the cruise report markdown process using the function get_biomass_and_nums_age_data_function.R; this provides all the data that is used to create a 'current survey only/pollock only' dataframe named current_survey_pollock_biomass_nums_age. This dataframe is the easiest input to the build_biomass_at_age_table_summer_goa function.

Value

A list with two items: item 1 is the Flextable table object, item 2 is the table caption.

Author

Mike Levine

Examples

if (FALSE) { # \dontrun{
# build the table and caption
biomass_at_age_table_list <-
  build_biomass_at_age_table_summer_goa
(biomass_nums_age_data <- current_survey_pollock_biomass_nums_age)
#
biomass_at_age_table <- biomass_at_age_table_list[[1]]
biomass_at_age_caption <- biomass_at_age_table_list[[2]]
} # }