Create biomass at age table for MACE winter cruise reports

Build the pollock biomass-at-age table used in winter GOA reports. The data used here is usually limited to the 'selectivity-corrected era' within Shelikof Strait (i.e. 2008-current). This function will report all ages from 1-18, and will do so for any region that was defined for the analysis. If you have an especially old age class (>18 years old), the table will report to this age as opposed to the nomimal age 1- age 18 range.

Usage

build_biomass_at_age_table_winter_goa(biomass_nums_age_data, region_name)

Arguments

biomass_nums_age_data

The name of the dataframe that contains all the abundance-at-age values; this dataframe is created in the cruise report markdown process using the function get_biomass_and_nums_age_data_function.R; this provides all the data that is then used to create a dataframe named shelikof_sel_corr_surveys_biomass_nums_age. This dataframe is the easiest input to build_biomass_at_age_table_winter_goa function.

region_name

The name of the reporting region that you want to include in this table (generally "Shelikof Strait" in winter GOA reports)

Value

A list with two items: item 1 is the Flextable table object, item 2 is the table caption.

Author

Mike Levine

Examples

if (FALSE) { # \dontrun{
# build the table and the caption
biomass_age_table_list <- build_numbers_at_age_table_winter_goa(
  biomass_nums_age_data = shelikof_sel_corr_surveys_biomass_nums_age,
  region_name = "Shelikof Strait"
)

# pull out the table and caption from the list
biomass_age_table <- biomass_age_table_list[[1]]
biomass_age_caption <- biomass_age_table_list[[2]]
} # }