An example dataset consisting of pollock length, weight, and age data from Gulf of Alasks summer MACE surveys

An example dataset consisting of pollock length, weight, and age data from Gulf of Alasks summer MACE surveys

Usage

pollock_length_weight_age_data

Format

A data frame with 19061 rows and 9 columns:

SURVEY

A unique identifier for each survey (numeric)

HAUL

An identifier for the trawl number that a specimen was captured in

SPECIES_CODE

An identifier for the species code for a given specimen

SUBCATEGORY

An identifier for the life history stage for a given specimen

FORK_LENGTH

pollock length (cm; numeric)

ORGANISM_WEIGHT

pollock weight (kg; numeric)

AGE

Pollock age (if available) (years; numeric)

REPORT_NUMBER

An identifier for the survey region in which a given specimen was captured (numeric)

region

An description of the survey region in which a given specimen was captured (character)

Source

Historic MACE reported values