shelikof_historic_numbers_by_age_1981_to_2007

values are stored as static, formatted 'canned' values gleaned from historic cruise reports. If you want to change the static dataframe (for example, removing years as more selectivity-corrected values become available, the data is available within MACEReports package: see shelikof_historic_numbers_by_age_1981_to_2007.

Usage

shelikof_historic_numbers_by_age_1981_to_2007

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 19 rows and 25 columns.

Source

Dataset initially compiled from historic cruise report tables