template_df_haul_data

A subset of standard MACE haul data, with columns and formats required for creating the standard haul table in cruise reports

Usage

template_df_haul_data

Format

A data frame with 5 rows and 15 columns:

EVENT_ID

Event number (numeric)

GEAR

Gear type (character)

EQ_TIME

The EQ time entered for the event ("POSIXct" "POSIXt" )

HB_TIME

The haulback time entered for the event ("POSIXct" "POSIXt" )

DURATION_MINS

Time from EQ-HB (numeric)

EQ_LATITUDE

Latitude (decimal degrees; numeric)

EQ_LONGITUDE

Longitdue (decimal degrees; numeric)

HEAD_ROPE_DEPTH

Mean headrope depth during event (numeric)

BOTTOM_DEPTH

Mean bottom depth during event (numeric)

HEAD_ROPE_TEMP

Mean head rope temperature during events (Deg C; numeric)

SURFACE_TEMP

Mean surface temperature during events (Deg C; numeric)

POLLOCK_WEIGHT

Walleye pollock catch weight (kg; numeric)

POLLOCK_NUMBERS

Walleye pollock catch numbers (individuals; numeric)

NON_POLLOCK_WEIGHT

All other species catch weight (kg; numeric)

region

Survey region name (character)

Source

Template dateset was gathered using get_haul_table_data.R function in Summer 2021 GOA Cruise Report